fix(moe): review pass on the multi-shard path
Three defects the split rewrite introduced, none of which the gates could see:

- The shard index rode in an int8_t, so a model past 127 shards wrapped to a
  negative index into the reader vector. The bounds check could never catch it:
  it validated the untruncated value. Widened to int16_t, which covers the whole
  -%05d-of-%05d filename space.
- DenseWeights::warm() reused one flag as both the inner loop condition and the
  partial-warm report, so the first shard that failed to open silently skipped
  the warm-up of every later shard. Per-shard condition, sticky report.
- The dense readers stayed allocated for the session after read_anonymous had
  copied and rebound every tensor: fds and a per-lane bounce buffer per shard,
  sitting next to a cache counting every MiB. Released at the end of init.

Also: the streaming banner read O_DIRECT off shard 0, which under the
small-first-shard layout is metadata only and too short to verify, so it could
claim a mode the shards carrying experts had not got. It now reports the weakest
of the readers.

+++ core/src/moe/expert_stream_source.h
@@ -114,7 +114,10 @@ private:
         uint64_t off = 0;
         uint64_t nbytes = 0;
         int32_t flag = -1; // overlap: index into ready_ to publish on completion; -1 = serial
-        int8_t file = 0;   // which shard reader serves this read
+        // Which shard reader serves this read. Wide enough for the whole filename format
+        // (-%05d-of-%05d): an int8_t would wrap a 200-shard model into a negative index that the
+        // init-time bounds check, which sees the untruncated value, could never catch.
+        int16_t file = 0;
         // Which (layer, expert, projection) this read serves. Known at every enqueue site and
         // otherwise thrown away; carried so the I/O trace can attribute a read without the read
         // path having to guess. Inert unless the trace is on.
